+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 - fix: svg renderer minimap hidden elements probem, closes: #2174;
 - feat: add extra parameter graph for menu plugin, closes: #2204;
 - fix: tooltip plugin, crossing different shape cant execute the getContent function, closes: #2153;
+- feat: add edgeConfig for create-edge behavior, closes: #2195;
+- fix: remove the source node while creat-edge;
+- feat: create-edge for combo, closes: #2211;
 
 #### 3.8.1
 - fix: update edge states with updateItem problem, closes: #2142;